+++ src/qdmi/devices/na/CMakeLists.txt
@@ -62,7 +62,8 @@ if(NOT TARGET ${TARGET_NAME})
   target_sources(${TARGET_NAME} PRIVATE App.cpp)
 
   # Link Generator library
-  target_link_libraries(${TARGET_NAME} PRIVATE MQT::CoreQDMINaDeviceGen spdlog::spdlog)
+  target_link_libraries(${TARGET_NAME} PRIVATE MQT::CoreQDMINaDeviceGen MQT::ProjectOptions
+                                               MQT::ProjectWarnings spdlog::spdlog)
 
   # set versioning information
   set_target_properties(${TARGET_NAME} PROPERTIES VERSION ${PROJECT_VERSION} EXPORT_NAME
@@ -133,7 +134,7 @@ if(NOT TARGET ${TARGET_NAME})
   target_link_libraries(
     ${TARGET_NAME}
     PUBLIC qdmi::qdmi
-    PRIVATE MQT::CoreQDMIBaseDevice MQT::ProjectOptions MQT::ProjectWarnings spdlog::spdlog)
+    PRIVATE MQT::CoreQDMIBaseDevice spdlog::spdlog)
 
   # Always compile with position independent code such that the library can be used in shared
   # libraries
