+++ src/qdmi/devices/na/Device.cpp
@@ -222,7 +222,13 @@ auto MQT_NA_QDMI_Device_Session_impl_d::queryDeviceProperty(
   if (status_ != Status::INITIALIZED) {
     return QDMI_ERROR_BADSTATE;
   }
-  return qdmi::na::Device::get()->queryProperty(prop, size, value, sizeRet);
+  try {
+    // Use the new get() that returns a shared_ptr.
+    return qdmi::na::Device::get()->queryProperty(prop, size, value, sizeRet);
+  } catch (const std::runtime_error&) {
+    // This happens if get() throws because the device is not initialized.
+    return QDMI_ERROR_BADSTATE;
+  }
 }
 auto MQT_NA_QDMI_Device_Session_impl_d::querySiteProperty(
     MQT_NA_QDMI_Site site, const QDMI_Site_Property prop, const size_t size,
@@ -634,7 +640,13 @@ int MQT_NA_QDMI_device_finalize() {
 }
 
 int MQT_NA_QDMI_device_session_alloc(MQT_NA_QDMI_Device_Session* session) {
-  return qdmi::na::Device::get()->sessionAlloc(session);
+  try {
+    // Use the new get() that returns a shared_ptr.
+    return qdmi::na::Device::get()->sessionAlloc(session);
+  } catch (const std::runtime_error&) {
+    // This happens if get() throws because the device is not initialized.
+    return QDMI_ERROR_BADSTATE;
+  }
 }
 
 int MQT_NA_QDMI_device_session_init(MQT_NA_QDMI_Device_Session session) {
@@ -645,7 +657,11 @@ int MQT_NA_QDMI_device_session_init(MQT_NA_QDMI_Device_Session session) {
 }
 
 void MQT_NA_QDMI_device_session_free(MQT_NA_QDMI_Device_Session session) {
-  qdmi::na::Device::get()->sessionFree(session);
+  try {
+    // Use the new get() that returns a shared_ptr.
+    return qdmi::na::Device::get()->sessionFree(session);
+  } catch (const std::runtime_error&) {
+  }
 }
 
 int MQT_NA_QDMI_device_session_set_parameter(
