+++ src/qdmi/devices/sc/Device.cpp
@@ -17,14 +17,18 @@
 #include "mqt_sc_qdmi/device.h"
 #include "qdmi/devices/sc/DeviceMemberInitializers.hpp"
 
+#include <algorithm>
 #include <atomic>
 #include <cassert>
 #include <cstddef>
 #include <cstdint>
 #include <cstring>
+#include <functional>
 #include <memory>
-#include <stdexcept>
+#include <span>
+#include <type_traits>
 #include <utility>
+#include <variant>
 #include <vector>
 
 // NOLINTBEGIN(bugprone-macro-parentheses)
@@ -101,6 +105,7 @@ Device::Device() {
   // NOLINTNEXTLINE(misc-const-correctness)
   INITIALIZE_SITES(sites_);
   INITIALIZE_COUPLINGMAP(couplingMap_);
+  INITIALIZE_OPERATIONS(operations_);
 }
 Device::~Device() {
   // Explicitly clear sessions before destruction to avoid spurious segfaults
@@ -154,6 +159,15 @@ auto Device::queryProperty(const QDMI_Device_Property prop, const size_t size,
                     couplingMap_, prop, size, value, sizeRet)
   ADD_LIST_PROPERTY(QDMI_DEVICE_PROPERTY_OPERATIONS, MQT_SC_QDMI_Operation,
                     operations_, prop, size, value, sizeRet)
+  if (prop == (QDMI_DEVICE_PROPERTY_SUPPORTEDPROGRAMFORMATS)) {
+    if (value != nullptr && size > 0) {
+      return QDMI_ERROR_INVALIDARGUMENT;
+    }
+    if (sizeRet != nullptr) {
+      *sizeRet = 0;
+    }
+    return QDMI_SUCCESS;
+  }
   return QDMI_ERROR_NOTSUPPORTED;
 }
 } // namespace qdmi::sc
@@ -303,21 +317,146 @@ auto MQT_SC_QDMI_Site_impl_d::queryProperty(const QDMI_Site_Property prop,
                             value, sizeRet)
   return QDMI_ERROR_NOTSUPPORTED;
 }
-auto MQT_SC_QDMI_Operation_impl_d::makeUnique()
+auto MQT_SC_QDMI_Operation_impl_d::sortSites() -> void {
+  std::visit(
+      [](auto& sites) {
+        using T = std::decay_t<decltype(sites)>;
+        if constexpr (std::is_same_v<T, std::vector<MQT_SC_QDMI_Site>>) {
+          // Single-qubit: sort flat list by pointer address
+          std::ranges::sort(sites, std::less<MQT_SC_QDMI_Site>{});
+        } else if constexpr (std::is_same_v<
+                                 T, std::vector<std::pair<MQT_SC_QDMI_Site,
+                                                          MQT_SC_QDMI_Site>>>) {
+          // Two-qubit: normalize each pair (first < second)
+          // Use std::less for proper total order (pointer comparison with
+          // operator> invokes undefined behavior)
+          std::ranges::for_each(sites, [](auto& p) {
+            if (std::less<MQT_SC_QDMI_Site>{}(p.second, p.first)) {
+              std::swap(p.first, p.second);
+            }
+          });
+          std::ranges::sort(sites);
+        }
+        // more cases go here if needed in the future
+      },
+      supportedSites_);
+}
+MQT_SC_QDMI_Operation_impl_d::MQT_SC_QDMI_Operation_impl_d(
+    std::string name, const size_t numParameters,
+    const std::vector<MQT_SC_QDMI_Site>& sites)
+    : name_(std::move(name)), numParameters_(numParameters), numQubits_(1),
+      supportedSites_(sites) {
+  sortSites();
+}
+MQT_SC_QDMI_Operation_impl_d::MQT_SC_QDMI_Operation_impl_d(
+    std::string name, const size_t numParameters,
+    const std::vector<std::pair<MQT_SC_QDMI_Site, MQT_SC_QDMI_Site>>& sites)
+    : name_(std::move(name)), numParameters_(numParameters), numQubits_(2),
+      supportedSites_(sites) {
+  sortSites();
+}
+auto MQT_SC_QDMI_Operation_impl_d::makeUniqueSingleQubit(
+    std::string name, const size_t numParameters,
+    const std::vector<MQT_SC_QDMI_Site>& sites)
     -> std::unique_ptr<MQT_SC_QDMI_Operation_impl_d> {
-  const MQT_SC_QDMI_Operation_impl_d op{};
+  const MQT_SC_QDMI_Operation_impl_d op(std::move(name), numParameters, sites);
   return std::make_unique<MQT_SC_QDMI_Operation_impl_d>(op);
 }
-// NOLINTNEXTLINE(readability-convert-member-functions-to-static)
+auto MQT_SC_QDMI_Operation_impl_d::makeUniqueTwoQubit(
+    std::string name, const size_t numParameters,
+    const std::vector<std::pair<MQT_SC_QDMI_Site, MQT_SC_QDMI_Site>>& sites)
+    -> std::unique_ptr<MQT_SC_QDMI_Operation_impl_d> {
+  const MQT_SC_QDMI_Operation_impl_d op(std::move(name), numParameters, sites);
+  return std::make_unique<MQT_SC_QDMI_Operation_impl_d>(op);
+}
 auto MQT_SC_QDMI_Operation_impl_d::queryProperty(
     const size_t numSites, const MQT_SC_QDMI_Site* sites,
     const size_t numParams, const double* params,
     const QDMI_Operation_Property prop, const size_t size, void* value,
-    size_t* /* unused */) const -> int {
+    size_t* sizeRet) const -> int {
   if ((sites != nullptr && numSites == 0) ||
       (params != nullptr && numParams == 0) ||
       (value != nullptr && size == 0) || prop >= QDMI_OPERATION_PROPERTY_MAX) {
     return QDMI_ERROR_INVALIDARGUMENT;
+  }
+  if (sites != nullptr) {
+    // If numQubits_ == 1 or isZoned_ == true
+    if (numSites == 1) {
+      // If the (single) site is not supported, return with an error
+      const bool found = std::visit(
+          [sites]<typename S>(const S& storedSites) -> bool {
+            using T = std::decay_t<S>;
+            if constexpr (std::is_same_v<T, std::vector<MQT_SC_QDMI_Site>>) {
+              return std::ranges::binary_search(storedSites, *sites,
+                                                std::less<MQT_SC_QDMI_Site>{});
+            }
+            return false; // Wrong variant type
+          },
+          supportedSites_);
+      if (!found) {
+        return QDMI_ERROR_NOTSUPPORTED;
+      }
+    } else if (numSites == 2) {
+      // NOLINTBEGIN(cppcoreguidelines-pro-bounds-pointer-arithmetic)
+      const std::pair needle = std::less<MQT_SC_QDMI_Site>{}(sites[0], sites[1])
+                                   ? std::make_pair(sites[0], sites[1])
+                                   : std::make_pair(sites[1], sites[0]);
+      // NOLINTEND(cppcoreguidelines-pro-bounds-pointer-arithmetic)
+      // if the pair of sites is not supported, return with an error
+      const bool found = std::visit(
+          [&needle]<typename S>(const S& storedSites) -> bool {
+            using T = std::decay_t<S>;
+            if constexpr (std::is_same_v<
+                              T, std::vector<std::pair<MQT_SC_QDMI_Site,
+                                                       MQT_SC_QDMI_Site>>>) {
+              return std::ranges::binary_search(storedSites, needle);
+            }
+            return false; // Wrong variant type
+          },
+          supportedSites_);
+      if (!found) {
+        return QDMI_ERROR_NOTSUPPORTED;
+      }
+    } // this device does not support operations with more than two qubits
+  }
+  ADD_STRING_PROPERTY(QDMI_OPERATION_PROPERTY_NAME, name_.c_str(), prop, size,
+                      value, sizeRet)
+  ADD_SINGLE_VALUE_PROPERTY(QDMI_OPERATION_PROPERTY_PARAMETERSNUM, size_t,
+                            numParameters_, prop, size, value, sizeRet)
+  ADD_SINGLE_VALUE_PROPERTY(QDMI_OPERATION_PROPERTY_QUBITSNUM, size_t,
+                            numQubits_, prop, size, value, sizeRet)
+  if (prop == QDMI_OPERATION_PROPERTY_SITES) {
+    return std::visit(
+        [&]<typename S>(const S& storedSites) -> int {
+          using T = std::decay_t<S>;
+          if constexpr (std::is_same_v<T, std::vector<MQT_SC_QDMI_Site>>) {
+            // Single-qubit: return flat array
+            ADD_LIST_PROPERTY(QDMI_OPERATION_PROPERTY_SITES, MQT_SC_QDMI_Site,
+                              storedSites, prop, size, value, sizeRet)
+          } else if constexpr (std::is_same_v<
+                                   T,
+                                   std::vector<std::pair<MQT_SC_QDMI_Site,
+                                                         MQT_SC_QDMI_Site>>>) {
+            // Ensure std::pair has standard layout and expected size
+            static_assert(std::is_standard_layout_v<
+                          std::pair<MQT_SC_QDMI_Site, MQT_SC_QDMI_Site>>);
+            static_assert(
+                sizeof(std::pair<MQT_SC_QDMI_Site, MQT_SC_QDMI_Site>) ==
+                2 * sizeof(MQT_SC_QDMI_Site));
+            // Two-qubit: reinterpret as flat array of sites using std::span
+            // std::pair has standard layout, so the memory layout of
+            // vector<pair<Site, Site>> is equivalent to Site[2*N]
+            const auto flatView = std::span<const MQT_SC_QDMI_Site>(
+                // NOLINTNEXTLINE(cppcoreguidelines-pro-type-reinterpret-cast)
+                reinterpret_cast<const MQT_SC_QDMI_Site*>(storedSites.data()),
+                storedSites.size() * 2);
+            ADD_LIST_PROPERTY(QDMI_OPERATION_PROPERTY_SITES, MQT_SC_QDMI_Site,
+                              flatView, prop, size, value, sizeRet)
+          }
+          // more cases go here if needed in the future
+          return QDMI_ERROR_NOTSUPPORTED;
+        },
+        supportedSites_);
   }
   return QDMI_ERROR_NOTSUPPORTED;
 }
@@ -333,7 +472,13 @@ int MQT_SC_QDMI_device_finalize() {
 }
 
 int MQT_SC_QDMI_device_session_alloc(MQT_SC_QDMI_Device_Session* session) {
-  return qdmi::sc::Device::get()->sessionAlloc(session);
+  try {
+    // Use the new get() that returns a shared_ptr.
+    return qdmi::sc::Device::get()->sessionAlloc(session);
+  } catch (const std::runtime_error&) {
+    // This happens if get() throws because the device is not initialized.
+    return QDMI_ERROR_BADSTATE;
+  }
 }
 
 int MQT_SC_QDMI_device_session_init(MQT_SC_QDMI_Device_Session session) {
@@ -344,7 +489,11 @@ int MQT_SC_QDMI_device_session_init(MQT_SC_QDMI_Device_Session session) {
 }
 
 void MQT_SC_QDMI_device_session_free(MQT_SC_QDMI_Device_Session session) {
-  qdmi::sc::Device::get()->sessionFree(session);
+  try {
+    // Use the new get() that returns a shared_ptr.
+    qdmi::sc::Device::get()->sessionFree(session);
+  } catch (const std::runtime_error&) {
+  }
 }
 
 int MQT_SC_QDMI_device_session_set_parameter(
